+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.1.3] - 2023-05-22
+
+## Added
+- LFO visualization (still missing some parameters)
+
+## Improved
+- Form Update
+- Accordion header toggle events and icon update
+
+
+
 # [0.1.2]
 
 ## Added